enhance code and remove unused ones

+++ src/state/plan/action-creators.js
@@ -1,70 +1,60 @@
 
 
 import _ from "lodash";
-
-
 import { AsyncStorage } from "react-native";
 import { arabicCheckedList, ArabicContent, checkedList, content } from '../../constants';
-import * as types from "./actions";
-
+import * as planActions from "./actions";
+//this function to load the plan content from local storage
 export async function loadPlan() {
-  let payloadPx =  JSON.parse(await AsyncStorage.getItem("plan"))
+  let payload = JSON.parse(await AsyncStorage.getItem("plan"))
   return {
-    type: types.LOAD_PLAN_CHAPTERS,
-    payload: payloadPx
+    type: planActions.LOAD_PLAN_CHAPTERS,
+    payload: payload
   }
 }
+
 export function selectDayOfPlan(dayNumber , isArabic = false){
   return async (dispatch, getState) => {
-    let planContent ;
+    let planContent;
     if(isArabic)
-    planContent =  getState().plan.ArabicPlanContent;
+    planContent = getState().plan.ArabicPlanContent;
     else
-    planContent =  getState().plan.planContent;
+    planContent = getState().plan.planContent;
    let content  = planContent[dayNumber].dayChapters
    let payload = {
      selectedDay:dayNumber ,
      selectedDayContent : content
    }
     dispatch({
-      type: types.SELECT_DAY_OF_PLAN,
+      type: planActions.SELECT_DAY_OF_PLAN,
       payload: payload
     });
   }
 }
 
-export async function makeChapterChecked (indexOfChapter) {
-  return async (dispatch, getState) => {
-    let checkedChaptersIndexes = getState().plan.checkedChaptersIndexes;
-    if(_.indexOf(checkedChaptersIndexes , indexOfChapter) === -1)
-    checkedChaptersIndexes.push(indexOfChapter);
-    else
-    checkedChaptersIndexes= _.filter(checkedChaptersIndexes , f => f !== indexOfChapter)
-
-  dispatch(
-    {
-      type: types.MAKE_CHAPETER_CHECKED,
-      payload: checkedChaptersIndexes
-    }
-  )
-}
-}
-
 export async function loadPlanCheckedList (){
  let res = await AsyncStorage.getItem("list")
- console.log('load checkedlist is' , res)
+//  console.log('load checkedlist is' , res)
   return {
-    type: types.LOAD_PLAN_CHECKED_LIST,
+    type: planActions.LOAD_PLAN_CHECKED_LIST,
     payload: JSON.parse(res) 
   }
 }
 
 export async function inializePlan(){
   return async dispatch => {
-      await AsyncStorage.setItem("plan" , JSON.stringify(content));
+      let isPlanContentExistOnLocalStorage = await AsyncStorage.getItem("plan");
+      if(!isPlanContentExistOnLocalStorage){
+        console.log("will set the plan content in local storage")
+        await AsyncStorage.setItem("plan" , JSON.stringify(content));
+      }
+      else 
+      {
+        console.log("no need to set the plan content in local storage")
+      }
   }
 }
-export async function inializeCheckedList(){
+export async function resetCheckedList(){
   return async dispatch => {
       await AsyncStorage.setItem("list" , JSON.stringify(checkedList));
       await AsyncStorage.setItem("arabicList" , JSON.stringify(arabicCheckedList));
@@ -90,13 +80,13 @@ export async function saveCheckedListIntoLocalStorage(list , isArabic){
 }
 export  function selectChapterOfDayPlan (IndexOfChapter) {
 return {
-  type: types.SELECT_CHAPTER_OF_DAY_PLAN,
+  type: planActions.SELECT_CHAPTER_OF_DAY_PLAN,
   payload:IndexOfChapter
 }
 }
 export function clearDayContentOfPlan(){
   return {
-    type: types.CLEAR_DAY_CONTENT_OF_PLAN
+    type: planActions.CLEAR_DAY_CONTENT_OF_PLAN
   }
 }
 
@@ -109,7 +99,7 @@ export async function inializeArabicPlan(){
 export async function loadArabicPlan() {
   let payload =  JSON.parse(await AsyncStorage.getItem("ArabicPlan"))
   return {
-    type: types.LOAD_ARABIC_PLAN_CHAPTERS,
+    type: planActions.LOAD_ARABIC_PLAN_CHAPTERS,
     payload: payload
   }
 }
@@ -117,10 +107,10 @@ export function setCompletedDayPlan(dayNumber){
   return async (dispatch, getState) => {
     let listOfCompletedDaysObj = getState().plan.listOfCompletedDaysObj;
     listOfCompletedDaysObj = {...listOfCompletedDaysObj,[dayNumber]:"true"};
-    console.log("before set" , listOfCompletedDaysObj)
+    // console.log("before set" , listOfCompletedDaysObj)
     await AsyncStorage.setItem("listOfCompletedDaysObj" , JSON.stringify(listOfCompletedDaysObj));
     dispatch({
-      type: types.SET_COMPLETED_DAY_PLAN_FLAG,
+      type: planActions.SET_COMPLETED_DAY_PLAN_FLAG,
       payload:listOfCompletedDaysObj
     })
  }
@@ -128,9 +118,9 @@ export function setCompletedDayPlan(dayNumber){
 export function loadListOfCompletedDaysOfPlan(){
  return async (dispatch, getState) => {
   let res =  await AsyncStorage.getItem("listOfCompletedDaysObj");
-  console.log("res" , res)
+  // console.log("res" , res)
    dispatch({
-    type: types.LOAD_LIST_OF_COMPLETED_DAYS_PLAN,
+    type: planActions.LOAD_LIST_OF_COMPLETED_DAYS_PLAN,
     payload: JSON.parse(res)  
    })
   }
